Moved to module approach Removed Scons output if not in debug Made Log a Singleton Moved to 1.1
