+++ Modules/__init__.py
@@ -1,0 +1,6 @@
+
+#import all necessary modules
+from Modules import Estragon_Config
+from Modules import Estragon_GodotEditor
+from Modules import Estragon_Log
+
